cambio 1 para API

+++ backend/src/main/java/ucb/edu/bo/SIS213/Cyberbaazar/CyberBaazar/dao/CountryDAO.java
@@ -3,6 +3,6 @@ package ucb.edu.bo.SIS213.Cyberbaazar.CyberBaazar.dao;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ucb.edu.bo.SIS213.Cyberbaazar.CyberBaazar.dto.CountryDTO;
 
-public interface CountryDAO extends JpaRepository<CountryDTO, Integer> {
+public interface CountryDAO extends JpaRepository<CountryDTO, Long> {
     
 }
