+++ backend/src/main/java/ucb/edu/bo/SIS213/Cyberbaazar/CyberBaazar/dao/EStoreDAO.java
@@ -3,7 +3,7 @@ package ucb.edu.bo.SIS213.Cyberbaazar.CyberBaazar.dao;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ucb.edu.bo.SIS213.Cyberbaazar.CyberBaazar.dto.EStoreDTO;
 
-public interface EStoreDAO extends JpaRepository<EStoreDTO, Integer> {
+public interface EStoreDAO extends JpaRepository<EStoreDTO, Long> {
     
 }
