+++ backend/src/main/java/ucb/edu/bo/SIS213/Cyberbaazar/CyberBaazar/dao/JobDAO.java
@@ -3,7 +3,7 @@ package ucb.edu.bo.SIS213.Cyberbaazar.CyberBaazar.dao;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ucb.edu.bo.SIS213.Cyberbaazar.CyberBaazar.dto.JobDTO;
 
-public interface JobDAO extends JpaRepository<JobDTO, Integer> {
+public interface JobDAO extends JpaRepository<JobDTO, Long> {
     
 }
