+++ backend/src/main/java/ucb/edu/bo/SIS213/Cyberbaazar/CyberBaazar/dao/ProductDAO.java
@@ -3,7 +3,7 @@ package ucb.edu.bo.SIS213.Cyberbaazar.CyberBaazar.dao;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ucb.edu.bo.SIS213.Cyberbaazar.CyberBaazar.dto.ProductDTO;
 
-public interface ProductDAO extends JpaRepository<ProductDTO, Integer> {
+public interface ProductDAO extends JpaRepository<ProductDTO, Long> {
     
 }
