+++ backend/src/main/java/ucb/edu/bo/SIS213/Cyberbaazar/CyberBaazar/dao/UserDAO.java
@@ -3,6 +3,6 @@ package ucb.edu.bo.SIS213.Cyberbaazar.CyberBaazar.dao;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ucb.edu.bo.SIS213.Cyberbaazar.CyberBaazar.dto.UserDTO;
 
-public interface UserDAO extends JpaRepository<UserDTO, Integer> {
+public interface UserDAO extends JpaRepository<UserDTO, Long> {
     
 }
